Split Drawing API into modules based on header files
