+++ src/drawing.rs
@@ -1,19 +1,13 @@
 #[link(name = "native_drawing")]
 extern "C" {}
 
-mod drawing_api10;
-pub use drawing_api10::*;
-
-#[cfg(feature = "api-11")]
-#[cfg_attr(docsrs, doc(cfg(feature = "api-11")))]
-mod api11_additions;
-#[cfg(feature = "api-11")]
-#[cfg_attr(docsrs, doc(cfg(feature = "api-11")))]
-pub use api11_additions::*;
-
-#[cfg(feature = "api-12")]
-#[cfg_attr(docsrs, doc(cfg(feature = "api-12")))]
-mod api12_additions;
-#[cfg(feature = "api-12")]
-#[cfg_attr(docsrs, doc(cfg(feature = "api-12")))]
-pub use api12_additions::*;
+pub mod bitmap;
+pub mod brush;
+pub mod canvas;
+pub mod color;
+pub mod font_collection;
+pub mod path;
+pub mod pen;
+pub mod text_declaration;
+pub mod text_typography;
+pub mod types;
