+++ src/drawing/bitmap.rs
@@ -1,0 +1,2 @@
+mod bitmap_api10;
+pub use bitmap_api10::*;
