+++ src/drawing/brush/brush_api10.rs
@@ -1,0 +1,58 @@
+/* automatically generated by rust-bindgen 0.69.4 */
+
+#![allow(non_upper_case_globals)]
+#![allow(non_camel_case_types)]
+#![allow(non_snake_case)]
+use crate::drawing::types::*;
+
+extern "C" {
+    /** @brief Creates an <b>OH_Drawing_Brush</b> object.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @return Returns the pointer to the <b>OH_Drawing_Brush</b> object created.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_BrushCreate() -> *mut OH_Drawing_Brush;
+    /** @brief Destroys an <b>OH_Drawing_Brush</b> object and reclaims the memory occupied by the object.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @param OH_Drawing_Brush Indicates the pointer to an <b>OH_Drawing_Brush</b> object.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_BrushDestroy(arg1: *mut OH_Drawing_Brush);
+    /** @brief Checks whether anti-aliasing is enabled for a brush. If anti-aliasing is enabled,
+    edges will be drawn with partial transparency.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @param OH_Drawing_Brush Indicates the pointer to an <b>OH_Drawing_Brush</b> object.
+    @return Returns <b>true</b> if anti-aliasing is enabled; returns <b>false</b> otherwise.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_BrushIsAntiAlias(arg1: *const OH_Drawing_Brush) -> bool;
+    /** @brief Enables or disables anti-aliasing for a brush. If anti-aliasing is enabled,
+    edges will be drawn with partial transparency.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @param OH_Drawing_Brush Indicates the pointer to an <b>OH_Drawing_Brush</b> object.
+    @param bool Specifies whether to enable anti-aliasing. The value <b>true</b> means
+                to enable anti-aliasing, and <b>false</b> means the opposite.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_BrushSetAntiAlias(arg1: *mut OH_Drawing_Brush, arg2: bool);
+    /** @brief Obtains the color of a brush. The color is used by the brush to fill in a shape.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @param OH_Drawing_Brush Indicates the pointer to an <b>OH_Drawing_Brush</b> object.
+    @return Returns a 32-bit (ARGB) variable that describes the color.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_BrushGetColor(arg1: *const OH_Drawing_Brush) -> u32;
+    /** @brief Sets the color for a brush. The color will be used by the brush to fill in a shape.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @param OH_Drawing_Brush Indicates the pointer to an <b>OH_Drawing_Brush</b> object.
+    @param color Indicates the color to set, which is a 32-bit (ARGB) variable.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_BrushSetColor(arg1: *mut OH_Drawing_Brush, color: u32);
+}
