+++ src/drawing/color.rs
@@ -1,0 +1,3 @@
+#[allow(unused_imports)]
+mod color_api10;
+pub use color_api10::*;
