+++ src/drawing/font_collection.rs
@@ -1,0 +1,3 @@
+#[allow(unused_imports)]
+mod font_collection_api10;
+pub use font_collection_api10::*;
