+++ src/drawing/font_collection/font_collection_api10.rs
@@ -1,0 +1,24 @@
+/* automatically generated by rust-bindgen 0.69.4 */
+
+#![allow(non_upper_case_globals)]
+#![allow(non_camel_case_types)]
+#![allow(non_snake_case)]
+use crate::drawing::text_declaration::*;
+use crate::drawing::types::*;
+
+extern "C" {
+    /** @brief Creates an <b>OH_Drawing_FontCollection</b> object.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @return Returns the pointer to the <b>OH_Drawing_FontCollection</b> object created.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_CreateFontCollection() -> *mut OH_Drawing_FontCollection;
+    /** @brief Releases the memory occupied by an <b>OH_Drawing_FontCollection</b> object.
+
+    @syscap SystemCapability.Graphic.Graphic2D.NativeDrawing
+    @param OH_Drawing_FontCollection Indicates the pointer to an <b>OH_Drawing_FontCollection</b> object.
+    @since 8
+    @version 1.0*/
+    pub fn OH_Drawing_DestroyFontCollection(arg1: *mut OH_Drawing_FontCollection);
+}
