+++ src/drawing/path.rs
@@ -1,0 +1,2 @@
+mod path_api10;
+pub use path_api10::*;
