+++ src/drawing/text_declaration.rs
@@ -1,0 +1,3 @@
+#[allow(unused_imports)]
+mod text_declaration_api10;
+pub use text_declaration_api10::*;
