+++ src/drawing/text_declaration/text_declaration_api10.rs
@@ -1,0 +1,27 @@
+/* automatically generated by rust-bindgen 0.69.4 */
+
+#![allow(non_upper_case_globals)]
+#![allow(non_camel_case_types)]
+#![allow(non_snake_case)]
+use crate::drawing::types::*;
+
+#[repr(C)]
+pub struct OH_Drawing_FontCollection {
+    _unused: [u8; 0],
+}
+#[repr(C)]
+pub struct OH_Drawing_Typography {
+    _unused: [u8; 0],
+}
+#[repr(C)]
+pub struct OH_Drawing_TextStyle {
+    _unused: [u8; 0],
+}
+#[repr(C)]
+pub struct OH_Drawing_TypographyStyle {
+    _unused: [u8; 0],
+}
+#[repr(C)]
+pub struct OH_Drawing_TypographyCreate {
+    _unused: [u8; 0],
+}
